// frontend/src/types/injury.ts
export interface Injury {
  id: number;
  wound_size: [number];
  summary_hemostasis: string;
  other_text: string;
  comment_injury_reason: string;
  injury_type: string;
  burn_percentage: [number];
  naca: number;
  bleeding_staunched_by_tourniquet: number;
  bleeding_staunched_by_bandage: number;
  bleeding_staunched_by_pelvic_belt: number;
  bleeding_staunched_by_quikclot: number;
  circulation_normal: number;
  circulation_none: number;
  circulation_reduced: number;
  motoric_normal: number;
  motoric_none: number;
  motoric_reduced: number;
  wound_open: number;
  wound_closed: number;
  blunt_trauma: number;
  penetrating_trauma: number;
  venous_bleeding: number;
  arterial_bleeding: number;
  spine_trauma_details_with_neurology: number;
  spine_trauma_details_without_neurology: number;
  amputation_detail_total: number;
  amputation_detail_subtotal: number;
  severity: number;
  location_head: number;
  location_extremities: number;
  location_throat: number;
  location_lower_torso: number;
  location_upper_torso: number;
  bandages_applied: number;
  pain_noted: number;
  ruled_out_fracture: number;
  injury_caused_by_fall: number;
  fracture: number;
  bruise: number;
  laceration: number;
  abrasion: number;
  traumatic_brain_injury: number;
  dislocation: number;
  trauma: number;
  amputation: number;
  other_injury_type: number;
  bite_wound: number;
  no_injury: number;
  burn_degree: number;
}